+++ clab/clab.go
@@ -15,7 +15,7 @@ import (
 
 	log "github.com/sirupsen/logrus"
 	"github.com/srl-labs/containerlab/cert"
-	"github.com/srl-labs/containerlab/cert/cfssl_ca"
+	"github.com/srl-labs/containerlab/cert/cfssl"
 	"github.com/srl-labs/containerlab/nodes"
 	"github.com/srl-labs/containerlab/runtime"
 	_ "github.com/srl-labs/containerlab/runtime/all"
@@ -146,7 +146,7 @@ func NewContainerLab(opts ...ClabOption) (*CLab, error) {
 
 		// init the Certificate Authority
 		c.certStorage = cert.NewLocalDiskCertStorage(c.TopoPaths)
-		c.rootCA = cfssl_ca.NewCertificatAuthorityCloudflare(c.certStorage, c.Config.Debug)
+		c.rootCA = cfssl.NewCertificatAuthority(c.certStorage, c.Config.Debug)
 	}
 	return c, err
 }
