+++ cmd/tools_cert.go
@@ -11,7 +11,7 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/srl-labs/containerlab/cert"
-	"github.com/srl-labs/containerlab/cert/cfssl_ca"
+	"github.com/srl-labs/containerlab/cert/cfssl"
 	"github.com/srl-labs/containerlab/utils"
 )
 
@@ -94,7 +94,7 @@ func createCA(_ *cobra.Command, _ []string) error {
 	log.Infof("Certificate attributes: CN=%s, C=%s, L=%s, O=%s, OU=%s, Validity period=%s",
 		commonName, country, locality, organization, organizationUnit, expiry)
 
-	rootCa := cfssl_ca.NewCertificatAuthorityCloudflare(nil, debug)
+	rootCa := cfssl.NewCertificatAuthority(nil, debug)
 
 	caCertInput := &cert.CsrInputCa{
 		CommonName:       commonName,
@@ -131,7 +131,7 @@ func signCert(_ *cobra.Command, _ []string) error {
 		}
 	}
 
-	rootCa := cfssl_ca.NewCertificatAuthorityCloudflare(nil, debug)
+	rootCa := cfssl.NewCertificatAuthority(nil, debug)
 
 	var caCert *cert.Certificate
 	log.Debugf("caCertPath: %q", caCertPath)
